feat: swap is ready
