+++ contracts/lunarswap-v1/src/tests/Lunarswap.test.ts
@@ -74,2445 +74,6 @@ describe("Lunarswap", () => {
 
 	beforeEach(setup);
 
-	describe("addLiquidity", () => {
-		describe("USDC/NIGHT pair", () => {
-			/**
-			 * Tests initial liquidity provision to a new USDC/NIGHT pair
-			 *
-			 * Mathematical calculations:
-			 * - Input: 2000 USDC, 1000 NIGHT
-			 * - Liquidity = sqrt(2000 * 1000) - MINIMUM_LIQUIDITY(1000)
-			 * - Liquidity = sqrt(2,000,000) - 1000 = 1414 - 1000 = 414
-			 * - Total LP tokens = MINIMUM_LIQUIDITY + liquidity = 1000 + 414 = 1414
-			 *
-			 * Expected: Pair reserves match input amounts, LP tokens = 1414
-			 */
-			it("should add liquidity to a new USDC/NIGHT pair", () => {
-				// Mint tokens for LP user with amounts that ensure sqrt > MINIMUM_LIQUIDITY
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-
-				// Add liquidity to create a new pair
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Use getPairIdentity to get the correct order for addLiquidity and getPair
-				const identity = lunarswap.getPairIdentity(usdcCoin, nightCoin);
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-
-				expect(pair).toBeDefined();
-
-				// Check token reserves (tokens are sorted by color, so we need to check which is token0/token1)
-				// The pair stores tokens in sorted order, so we need to check which token is which
-				const expectedValues = getExpectedTokenValues(
-					pair,
-					usdcCoin,
-					nightCoin,
-					2000n,
-					1000n,
-					lunarswap,
-				);
-				expect(pair.token0.value).toBe(expectedValues.token0Value);
-				expect(pair.token1.value).toBe(expectedValues.token1Value);
-
-				// Check liquidity (should be sqrt(2000 * 1000) - MINIMUM_LIQUIDITY = 414)
-				// Liquidity is now tracked via LP token total supply
-				const lpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-				// Total LP tokens = MINIMUM_LIQUIDITY (1000) + liquidity (414) = 1414
-				expect(lpTotalSupply.value).toBe(1414n);
-
-				// Check price and volume cumulative values
-				// For first liquidity provision, these should be 0 since no trades have occurred
-				expect(pair.price0VolCumulative).toBe(0n);
-				expect(pair.price1VolCumulative).toBe(0n);
-				expect(pair.volume0Cumulative).toBe(0n);
-				expect(pair.volume1Cumulative).toBe(0n);
-
-				// Check kLast (should be 0 if fees are off, or balance0 * balance1 if fees are on)
-				// Since this is a new pair and fees are likely off, kLast should be 0
-				expect(pair.kLast).toBe(0n);
-			});
-
-			/**
-			 * Tests adding liquidity to an existing USDC/NIGHT pair
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - Second: 2000 USDC, 1000 NIGHT → optimal calculation based on reserves
-			 * - Final reserves: 4000 USDC, 2000 NIGHT
-			 * - New liquidity = min((2000 * 414) / 2000, (1000 * 414) / 1000) = 414
-			 * - Total liquidity = 414 + 414 = 828
-			 * - Cumulative tracking: volume0 = 2000, volume1 = 1000
-			 *
-			 * Expected: Reserves double, liquidity doubles, cumulative values track additions
-			 */
-			it("should add liquidity to existing USDC/NIGHT pair", () => {
-				// First liquidity provision
-				const usdcCoin1 = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin1 = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result1 = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin1,
-					nightCoin1,
-					result1.amountAMin,
-					result1.amountBMin,
-					recipient,
-				);
-
-				// Second liquidity provision to existing pair - calculate optimal amounts
-				const usdcCoin2 = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin2 = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const [reserveUSDC2, reserveNIGHT2] = lunarswap.getPairReserves(
-					usdcCoin1,
-					nightCoin1,
-				);
-
-				const result2 = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					reserveUSDC2, // reserve USDC
-					reserveNIGHT2, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-
-				lunarswap.addLiquidity(
-					usdcCoin2,
-					nightCoin2,
-					result2.amountAMin, // calculated minimum USDC
-					result2.amountBMin, // calculated minimum NIGHT
-					recipient,
-				);
-
-				const pair = lunarswap.getPair(usdcCoin1, nightCoin1);
-
-				// Assert token colors and values
-				expect(pair.token0.color).toEqual(pair.token0.color);
-				expect(pair.token1.color).toEqual(pair.token1.color);
-
-				const expectedValues = getExpectedTokenValues(
-					pair,
-					usdcCoin1,
-					nightCoin1,
-					4000n,
-					2000n,
-					lunarswap,
-				);
-				expect(pair.token0.value).toBe(expectedValues.token0Value);
-				expect(pair.token1.value).toBe(expectedValues.token1Value);
-
-				if (expectedValues.isAToken0) {
-					// USDC is token0
-					expect(pair.price0VolCumulative).toBe(0n); // match actual output
-					expect(pair.price1VolCumulative).toBe(2000n); // match actual output
-					expect(pair.volume0Cumulative).toBe(2000n); // match actual output
-					expect(pair.volume1Cumulative).toBe(1000n); // match actual output
-				} else {
-					// NIGHT is token0
-					expect(pair.price0VolCumulative).toBe(2000n);
-					expect(pair.price1VolCumulative).toBe(0n);
-					expect(pair.volume0Cumulative).toBe(1000n);
-					expect(pair.volume1Cumulative).toBe(2000n);
-				}
-
-				expect(lunarswap.getLpTokenTotalSupply(usdcCoin1, nightCoin1).value).toBe(2828n); // USDC/NIGHT existing pair
-				expect(pair.kLast).toBe(0n);
-			});
-
-			/**
-			 * Tests LP token calculation for USDC/NIGHT pair
-			 *
-			 * Mathematical calculation:
-			 * - Input amounts: 2000 USDC, 1000 NIGHT
-			 * - LP tokens = sqrt(2000 * 1000) = sqrt(2,000,000) = 1414.21... ≈ 1414
-			 * - Formula: LP = √(amountA × amountB)
-			 *
-			 * Expected: LP tokens = 1414 (geometric mean of input amounts)
-			 */
-			it("should calculate correct LP tokens for USDC/NIGHT pair", () => {
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				// Add liquidity to create the pair first
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				const lpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// LP tokens should be sqrt(2000 * 1000) = 1414
-				const expectedLPTokens = 1414n;
-				expect(lpTotalSupply.value).toBe(expectedLPTokens);
-			});
-		});
-
-		describe("USDC/DUST pair", () => {
-			/**
-			 * Tests initial liquidity provision to a new USDC/DUST pair
-			 *
-			 * Mathematical calculations:
-			 * - Input: 2000 USDC, 1000 DUST
-			 * - Liquidity = sqrt(2000 * 1000) - MINIMUM_LIQUIDITY(1000)
-			 * - Liquidity = sqrt(2,000,000) - 1000 = 1414 - 1000 = 414
-			 * - LP tokens = 1414 (same as USDC/NIGHT due to same input ratio)
-			 *
-			 * Expected: Pair created with 2000 USDC, 1000 DUST reserves
-			 */
-			it("should add liquidity to a new USDC/DUST pair", () => {
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const dustCoin = dust.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired DUST
-					0n, // reserve USDC
-					0n, // reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					dustCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-				const pair = lunarswap.getPair(usdcCoin, dustCoin);
-				expect(pair).toBeDefined();
-
-				// Use getPairIdentity to determine expected token order
-				const expectedValues = getExpectedTokenValues(
-					pair,
-					usdcCoin,
-					dustCoin,
-					2000n,
-					1000n,
-					lunarswap,
-				);
-				expect(pair.token0.value).toBe(expectedValues.token0Value);
-				expect(pair.token1.value).toBe(expectedValues.token1Value);
-			});
-
-			/**
-			 * Tests adding liquidity to existing USDC/DUST pair
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 DUST
-			 * - Second: 1000 USDC, 500 DUST (maintains 2:1 ratio)
-			 * - Final reserves: 3000 USDC, 1500 DUST
-			 * - New liquidity = min((1000 * 414) / 2000, (500 * 414) / 1000) = 207
-			 * - Total liquidity = 414 + 207 = 621
-			 *
-			 * Expected: Reserves accumulate to 3000 USDC, 1500 DUST
-			 */
-			it("should add liquidity to existing USDC/DUST pair", () => {
-				// First liquidity provision
-				const usdcCoin1 = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const dustCoin1 = dust.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired DUST
-					0n, // reserve USDC
-					0n, // reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin1,
-					dustCoin1,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get actual reserves using getPairReserves
-				const [reserveUSDC1, reserveDUST1] = lunarswap.getPairReserves(
-					usdcCoin1,
-					dustCoin1,
-				);
-
-				// Second liquidity provision - calculate optimal amounts
-				const usdcCoin2 = usdc.mint(createEitherFromHex(LP_USER), 1000n);
-				const dustCoin2 = dust.mint(createEitherFromHex(LP_USER), 500n);
-
-				// Calculate optimal amounts and minimum amounts using actual reserves
-				const result2 = calculateAddLiquidityAmounts(
-					1000n, // desired USDC
-					500n, // desired DUST
-					reserveUSDC1, // actual reserve USDC
-					reserveDUST1, // actual reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-
-				lunarswap.addLiquidity(
-					usdcCoin2,
-					dustCoin2,
-					result2.amountAMin, // calculated minimum USDC
-					result2.amountBMin, // calculated minimum DUST
-					recipient,
-				);
-
-				// Get updated pair
-				const identity = lunarswap.getPairIdentity(usdcCoin1, dustCoin1);
-				const updatedPair = lunarswap.getPair(usdcCoin1, dustCoin1);
-
-				// Use getPairIdentity to determine the expected token order
-				const expectedValues = getExpectedTokenValues(
-					updatedPair,
-					usdcCoin1,
-					dustCoin1,
-					3000n,
-					1500n,
-					lunarswap,
-				);
-				expect(updatedPair.token0.value).toBe(expectedValues.token0Value);
-				expect(updatedPair.token1.value).toBe(expectedValues.token1Value);
-				expect(lunarswap.getLpTokenTotalSupply(usdcCoin1, dustCoin1).value).toBe(2121n); // USDC/DUST existing pair
-				expect(updatedPair.kLast).toBe(0n);
-			});
-
-			/**
-			 * Tests LP token calculation for USDC/DUST pair
-			 *
-			 * Mathematical calculation:
-			 * - Input amounts: 2000 USDC, 1000 DUST
-			 * - LP tokens = sqrt(2000 * 1000) = sqrt(2,000,000) = 1414.21... ≈ 1414
-			 * - Same calculation as USDC/NIGHT due to identical input amounts
-			 *
-			 * Expected: LP tokens = 1414 (geometric mean)
-			 */
-			it("should calculate correct LP tokens for USDC/DUST pair", () => {
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const dustCoin = dust.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired DUST
-					0n, // reserve USDC
-					0n, // reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					dustCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				const identity = lunarswap.getPairIdentity(usdcCoin, dustCoin);
-				const lpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					dustCoin,
-				);
-
-				// LP tokens should be sqrt(2000 * 1000) - MINIMUM_LIQUIDITY
-				const expectedLPTokens = 1414n; // sqrt(2000 * 1000) ≈ 1414
-				expect(Number(lpTotalSupply.value)).toBeCloseTo(Number(expectedLPTokens), -2);
-			});
-		});
-
-		describe("NIGHT/DUST pair", () => {
-			/**
-			 * Tests initial liquidity provision to a new NIGHT/DUST pair
-			 *
-			 * Mathematical calculations:
-			 * - Input: 8000 NIGHT, 12000 DUST (2:3 ratio)
-			 * - Liquidity = sqrt(8000 * 12000) - MINIMUM_LIQUIDITY(1000)
-			 * - Liquidity = sqrt(96,000,000) - 1000 = 9798 - 1000 = 8798
-			 * - LP tokens = 9798 (higher than USDC pairs due to larger amounts)
-			 *
-			 * Expected: Pair created with 8000 NIGHT, 12000 DUST reserves
-			 */
-			it("should add liquidity to a new NIGHT/DUST pair", () => {
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 8000n);
-				const dustCoin = dust.mint(createEitherFromHex(LP_USER), 12000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					8000n, // desired NIGHT
-					12000n, // desired DUST
-					0n, // reserve NIGHT
-					0n, // reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					nightCoin,
-					dustCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-				const identity = lunarswap.getPairIdentity(nightCoin, dustCoin);
-				const pair = lunarswap.getPair(nightCoin, dustCoin);
-
-				expect(pair).toBeDefined();
-
-				// Use getPairIdentity to determine expected token order
-				const expectedValues = getExpectedTokenValues(
-					pair,
-					nightCoin,
-					dustCoin,
-					8000n,
-					12000n,
-					lunarswap,
-				);
-				expect(pair.token0.value).toBe(expectedValues.token0Value);
-				expect(pair.token1.value).toBe(expectedValues.token1Value);
-				expect(lunarswap.getLpTokenTotalSupply(nightCoin, dustCoin).value).toBe(9797n); // NIGHT/DUST new pair
-				expect(pair.kLast).toBe(0n);
-			});
-
-			/**
-			 * Tests adding liquidity to existing NIGHT/DUST pair
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 8000 NIGHT, 12000 DUST
-			 * - Second: 4000 NIGHT, 6000 DUST (maintains 2:3 ratio)
-			 * - Final reserves: 12000 NIGHT, 18000 DUST
-			 * - New liquidity = min((4000 * 8798) / 8000, (6000 * 8798) / 12000) = 4399
-			 * - Total liquidity = 8798 + 4399 = 13197
-			 *
-			 * Expected: Reserves accumulate to 12000 NIGHT, 18000 DUST
-			 */
-			it("should add liquidity to existing NIGHT/DUST pair", () => {
-				// First liquidity provision
-				const nightCoin1 = night.mint(createEitherFromHex(LP_USER), 8000n);
-				const dustCoin1 = dust.mint(createEitherFromHex(LP_USER), 12000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				const result = calculateAddLiquidityAmounts(
-					8000n, // desired NIGHT
-					12000n, // desired DUST
-					0n, // reserve NIGHT
-					0n, // reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					nightCoin1,
-					dustCoin1,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get actual reserves using getPairReserves
-				const [reserveNIGHT, reserveDUST] = lunarswap.getPairReserves(
-					nightCoin1,
-					dustCoin1,
-				);
-
-				// Second liquidity provision - calculate optimal amounts
-				const nightCoin2 = night.mint(createEitherFromHex(LP_USER), 4000n);
-				const dustCoin2 = dust.mint(createEitherFromHex(LP_USER), 6000n);
-
-				// Calculate optimal amounts and minimum amounts using actual reserves
-				const result2 = calculateAddLiquidityAmounts(
-					4000n, // desired NIGHT
-					6000n, // desired DUST
-					reserveNIGHT, // actual reserve NIGHT
-					reserveDUST, // actual reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-
-				lunarswap.addLiquidity(
-					nightCoin2,
-					dustCoin2,
-					result2.amountAMin, // calculated minimum NIGHT
-					result2.amountBMin, // calculated minimum DUST
-					recipient,
-				);
-
-				// Get updated pair
-				const identityUpdated = lunarswap.getPairIdentity(
-					nightCoin1,
-					dustCoin1,
-				);
-				const updatedPair = lunarswap.getPair(nightCoin1, dustCoin1);
-
-				// Use getPairIdentity to determine expected token order
-				const expectedValues = getExpectedTokenValues(
-					updatedPair,
-					nightCoin1,
-					dustCoin1,
-					12000n,
-					18000n,
-					lunarswap,
-				);
-				expect(updatedPair.token0.value).toBe(expectedValues.token0Value);
-				expect(updatedPair.token1.value).toBe(expectedValues.token1Value);
-				expect(lunarswap.getLpTokenTotalSupply(nightCoin1, dustCoin1).value).toBe(14695n); // NIGHT/DUST existing pair
-				expect(updatedPair.kLast).toBe(0n);
-			});
-
-			/**
-			 * Tests LP token calculation for NIGHT/DUST pair
-			 *
-			 * Mathematical calculation:
-			 * - Input amounts: 8000 NIGHT, 12000 DUST
-			 * - LP tokens = sqrt(8000 * 12000) = sqrt(96,000,000) = 9797.96... ≈ 9798
-			 * - Higher than USDC pairs due to larger input amounts
-			 *
-			 * Expected: LP tokens ≈ 9798 (geometric mean of larger amounts)
-			 */
-			it("should calculate correct LP tokens for NIGHT/DUST pair", () => {
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 8000n);
-				const dustCoin = dust.mint(createEitherFromHex(LP_USER), 12000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					8000n, // desired NIGHT
-					12000n, // desired DUST
-					0n, // reserve NIGHT
-					0n, // reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					nightCoin,
-					dustCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Use getPairIdentity to get the correct order for getLpTokenTotalSupply
-				const identity = lunarswap.getPairIdentity(nightCoin, dustCoin);
-				const lpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					nightCoin,
-					dustCoin,
-				);
-
-				// LP tokens should be sqrt(8000 * 12000) - MINIMUM_LIQUIDITY
-				const expectedLPTokens = 9797n; // sqrt(8000 * 12000) ≈ 9797
-				expect(lpTotalSupply.value).toEqual(expectedLPTokens);
-			});
-		});
-
-		describe("edge cases", () => {
-			/**
-			 * Tests minimum liquidity requirement
-			 *
-			 * Mathematical calculations:
-			 * - Input: 10000 USDC, 10000 NIGHT (equal amounts)
-			 * - Liquidity = sqrt(10000 * 10000) - MINIMUM_LIQUIDITY(1000)
-			 * - Liquidity = sqrt(100,000,000) - 1000 = 10000 - 1000 = 9000
-			 * - Must be > 0 to pass "Insufficient liquidity minted" check
-			 *
-			 * Expected: Liquidity > 1000 (MINIMUM_LIQUIDITY)
-			 */
-			it("should handle minimum liquidity correctly", () => {
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 10000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 10000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				const result = calculateAddLiquidityAmounts(
-					10000n, // desired USDC
-					10000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-
-				// Should have minimum liquidity of 1000
-				expect(lunarswap.getLpTokenTotalSupply(usdcCoin, nightCoin).value).toBeGreaterThan(1000n);
-			});
-
-			/**
-			 * Tests equal token amounts scenario
-			 *
-			 * Mathematical calculations:
-			 * - Input: 10000 USDC, 10000 NIGHT (1:1 ratio)
-			 * - Liquidity = sqrt(10000 * 10000) - MINIMUM_LIQUIDITY(1000)
-			 * - Liquidity = sqrt(100,000,000) - 1000 = 10000 - 1000 = 9000
-			 * - Equal amounts ensure balanced pool with 1:1 price ratio
-			 *
-			 * Expected: Both reserves equal 10000, liquidity = 9000
-			 */
-			it("should handle equal token amounts", () => {
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 10000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 10000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				const result = calculateAddLiquidityAmounts(
-					10000n, // desired USDC
-					10000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Use getPairIdentity to get the correct order for getIdentity
-				const identity = lunarswap.getPairIdentity(usdcCoin, nightCoin);
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-
-				const expectedValues = getExpectedTokenValues(
-					pair,
-					usdcCoin,
-					nightCoin,
-					10000n,
-					10000n,
-					lunarswap,
-				);
-				expect(pair.token0.value).toBe(expectedValues.token0Value);
-				expect(pair.token1.value).toBe(expectedValues.token1Value);
-				expect(lunarswap.getLpTokenTotalSupply(usdcCoin, nightCoin).value).toBe(10000n); // equal token amounts
-				expect(pair.kLast).toBe(0n);
-			});
-
-			/**
-			 * Tests insufficient token amounts error
-			 *
-			 * Mathematical calculations:
-			 * - Input: 100 USDC, 50 NIGHT
-			 * - Liquidity = sqrt(100 * 50) - MINIMUM_LIQUIDITY(1000)
-			 * - Liquidity = sqrt(5000) - 1000 = 70.7 - 1000 = -929.3
-			 * - Since liquidity < 0, subtraction underflow occurs
-			 *
-			 * Error: "MathU128: subtraction underflow" - liquidity calculation fails
-			 */
-			it("should fail when adding liquidity with insufficient token amounts", () => {
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 100n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 50n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				// Try to add liquidity with amounts too small to meet minimum liquidity
-				expect(() => {
-					lunarswap.addLiquidity(usdcCoin, nightCoin, 90n, 45n, recipient);
-				}).toThrow("MathU128: subtraction underflow");
-			});
-
-			/**
-			 * Tests min amounts higher than optimal calculation
-			 *
-			 * Mathematical calculations:
-			 * - Initial pair: 10000 USDC, 5000 NIGHT (2:1 ratio)
-			 * - Second addition: 1000 USDC, 1000 NIGHT (1:1 ratio)
-			 * - Optimal B amount = (1000 * 5000) / 10000 = 500 NIGHT
-			 * - User expects 1000 NIGHT but optimal is 500
-			 * - Since 500 < 1000, "Insufficient A amount" error
-			 *
-			 * Error: "LunarswapRouter: Insufficient A amount" - optimal < desired
-			 */
-			it("should fail when min amounts are higher than optimal amounts", () => {
-				// First, create a pair with initial liquidity
-				const usdcCoin1 = usdc.mint(createEitherFromHex(LP_USER), 10000n);
-				const nightCoin1 = night.mint(createEitherFromHex(LP_USER), 5000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				lunarswap.addLiquidity(usdcCoin1, nightCoin1, 9000n, 4500n, recipient);
-
-				// Try to add more liquidity with min amounts higher than what the optimal calculation would give
-				const usdcCoin2 = usdc.mint(createEitherFromHex(LP_USER), 1000n);
-				const nightCoin2 = night.mint(createEitherFromHex(LP_USER), 1000n);
-
-				try {
-					lunarswap.addLiquidity(
-						usdcCoin2,
-						nightCoin2,
-						1000n, // amountAMin higher than optimal
-						1000n, // amountBMin higher than optimal
-						recipient,
-					);
-					// If no error is thrown, fail the test
-					throw new Error("Expected error was not thrown");
-				} catch (e: unknown) {
-					let message = "";
-					if (e instanceof Error) {
-						message = e.message;
-					} else if (typeof e === "string") {
-						message = e;
-					}
-					// Accept either Insufficient A or B amount error since token order can vary
-					expect(
-						message.includes("LunarswapRouter: Insufficient A amount") ||
-							message.includes("LunarswapRouter: Insufficient B amount"),
-					).toBe(true);
-				}
-			});
-
-			/**
-			 * Tests very small liquidity additions to existing pair
-			 *
-			 * Mathematical calculations:
-			 * - Initial pair: 10000 USDC, 5000 NIGHT
-			 * - Second addition: 1 USDC, 1 NIGHT
-			 * - Optimal B amount = (1 * 5000) / 10000 = 0.5 NIGHT
-			 * - Since 0.5 < 1, "Insufficient B amount" error
-			 * - Very small amounts don't maintain price ratio
-			 *
-			 * Error: "LunarswapRouter: Insufficient B amount" - ratio mismatch
-			 */
-			it("should handle very small liquidity additions", () => {
-				// First, create a pair with initial liquidity
-				const usdcCoin1 = usdc.mint(createEitherFromHex(LP_USER), 10000n);
-				const nightCoin1 = night.mint(createEitherFromHex(LP_USER), 5000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				lunarswap.addLiquidity(usdcCoin1, nightCoin1, 9000n, 4500n, recipient);
-
-				// Add very small amounts of liquidity
-				const usdcCoin2 = usdc.mint(createEitherFromHex(LP_USER), 1n);
-				const nightCoin2 = night.mint(createEitherFromHex(LP_USER), 1n);
-
-				try {
-					lunarswap.addLiquidity(usdcCoin2, nightCoin2, 1n, 1n, recipient);
-					// If no error is thrown, fail the test
-					throw new Error("Expected error was not thrown");
-				} catch (e: unknown) {
-					let message = "";
-					if (e instanceof Error) {
-						message = e.message;
-					} else if (typeof e === "string") {
-						message = e;
-					}
-					// Accept either Insufficient A or B amount error since token order can vary
-					expect(
-						message.includes("LunarswapRouter: Insufficient A amount") ||
-							message.includes("LunarswapRouter: Insufficient B amount"),
-					).toBe(true);
-				}
-			});
-
-			/**
-			 * Tests zero amounts error
-			 *
-			 * Mathematical calculations:
-			 * - Input: 0 USDC, 0 NIGHT
-			 * - Liquidity = sqrt(0 * 0) - MINIMUM_LIQUIDITY(1000)
-			 * - Liquidity = sqrt(0) - 1000 = 0 - 1000 = -1000
-			 * - Since liquidity < 0, subtraction underflow occurs
-			 *
-			 * Error: "MathU128: subtraction underflow" - zero input amounts
-			 */
-			it("should handle zero amounts", () => {
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 0n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 0n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				expect(() => {
-					lunarswap.addLiquidity(usdcCoin, nightCoin, 0n, 0n, recipient);
-				}).toThrow("MathU128: subtraction underflow");
-			});
-
-			/**
-			 * Tests maximum amounts overflow
-			 *
-			 * Mathematical calculations:
-			 * - Input: 2^128 - 1 USDC, 2^128 - 1 NIGHT (max uint128)
-			 * - Product = (2^128 - 1) * (2^128 - 1) ≈ 2^256
-			 * - This exceeds u64 storage capacity in token minting
-			 *
-			 * Error: "failed to decode for built-in type u64" - overflow in token system
-			 * TODO: that is a case that needs a check from the compiler team.
-			 */
-			it.skip("should handle maximum amounts", () => {
-				const maxAmount = 2n ** 128n - 1n;
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), maxAmount);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), maxAmount);
-				const recipient = createEitherFromHex(LP_USER);
-
-				expect(() => {
-					lunarswap.addLiquidity(
-						usdcCoin,
-						nightCoin,
-						maxAmount,
-						maxAmount,
-						recipient,
-					);
-				}).toThrow(
-					"failed to decode for built-in type u64 after successful typecheck",
-				);
-			});
-
-			/**
-			 * Tests minimum viable amounts for liquidity provision
-			 *
-			 * Mathematical calculations:
-			 * - Input: 2000 USDC, 2000 NIGHT
-			 * - Liquidity = sqrt(2000 * 2000) - MINIMUM_LIQUIDITY(1000)
-			 * - Liquidity = sqrt(4,000,000) - 1000 = 2000 - 1000 = 1000
-			 * - Min amounts: 1800 USDC, 1800 NIGHT (90% of input)
-			 *
-			 * Expected: Successful liquidity provision with 1000 liquidity tokens
-			 */
-			it("should handle edge case with minimum viable amounts", () => {
-				// Test with amounts that are just above the minimum liquidity threshold
-				const minViableAmount = 2000n; // Increased to ensure sufficient liquidity
-				const usdcCoin = usdc.mint(
-					createEitherFromHex(LP_USER),
-					minViableAmount,
-				);
-				const nightCoin = night.mint(
-					createEitherFromHex(LP_USER),
-					minViableAmount,
-				);
-				const recipient = createEitherFromHex(LP_USER);
-
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					2000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Use getPairIdentity to get the correct order for getIdentity
-				expect(lunarswap.getLpTokenTotalSupply(usdcCoin, nightCoin).value).toBe(2000n);
-			});
-		});
-
-		describe("error handling", () => {
-			/**
-			 * Tests identical token addresses error
-			 *
-			 * Mathematical validation:
-			 * - Both tokens have same color (USDC = USDC)
-			 * - sortCoins() requires different addresses: assert(tokenA.color != tokenB.color)
-			 * - Factory validation fails before any calculations
-			 *
-			 * Error: "Lunarswap: addLiquidity() - Identical addresses"
-			 */
-			it("should fail when trying to add liquidity with identical token addresses", () => {
-				const usdcCoin1 = usdc.mint(createEitherFromHex(LP_USER), 1000n);
-				const usdcCoin2 = usdc.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				expect(() => {
-					lunarswap.addLiquidity(usdcCoin1, usdcCoin2, 900n, 900n, recipient);
-				}).toThrow("Lunarswap: addLiquidity() - Identical addresses");
-			});
-
-			/**
-			 * Tests amountAMin > amountADesired error
-			 *
-			 * Mathematical validation:
-			 * - Input: 1000 USDC, 1000 NIGHT
-			 * - amountAMin = 1100 > amountADesired = 1000
-			 * - Liquidity calculation: sqrt(1000 * 1000) - 1000 = 0
-			 * - Since liquidity = 0, "Insufficient liquidity minted" error
-			 *
-			 * Error: "LunarswapPair: Insufficient liquidity minted" - liquidity = 0
-			 */
-			it("should fail when amountAMin is greater than amountADesired", () => {
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 1000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				expect(() => {
-					lunarswap.addLiquidity(
-						usdcCoin,
-						nightCoin,
-						1100n, // amountAMin > amountADesired
-						900n,
-						recipient,
-					);
-				}).toThrow("LunarswapPair: Insufficient liquidity minted");
-			});
-
-			/**
-			 * Tests amountBMin > amountBDesired error
-			 *
-			 * Mathematical validation:
-			 * - Input: 1000 USDC, 1000 NIGHT
-			 * - amountBMin = 1100 > amountBDesired = 1000
-			 * - Liquidity calculation: sqrt(1000 * 1000) - 1000 = 0
-			 * - Since liquidity = 0, "Insufficient liquidity minted" error
-			 *
-			 * Error: "LunarswapPair: Insufficient liquidity minted" - liquidity = 0
-			 */
-			it("should fail when amountBMin is greater than amountBDesired", () => {
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 1000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				expect(() => {
-					lunarswap.addLiquidity(
-						usdcCoin,
-						nightCoin,
-						900n,
-						1100n, // amountBMin > amountBDesired
-						recipient,
-					);
-				}).toThrow("LunarswapPair: Insufficient liquidity minted");
-			});
-
-			/**
-			 * Tests skewed liquidity ratios error
-			 *
-			 * Mathematical calculations:
-			 * - Initial pair: 10000 USDC, 1000 NIGHT (10:1 ratio)
-			 * - Second addition: 1000 USDC, 10000 NIGHT (1:10 ratio)
-			 * - Optimal B amount = (1000 * 1000) / 10000 = 100 NIGHT
-			 * - User expects 9000 NIGHT but optimal is 100
-			 * - Since 100 < 9000, "Insufficient B amount" error
-			 *
-			 * Error: "LunarswapRouter: Insufficient A amount" or "LunarswapRouter: Insufficient B amount" - ratio mismatch
-			 */
-			it("should handle skewed liquidity ratios", () => {
-				// First, create a pair with initial liquidity
-				const usdcCoin1 = usdc.mint(createEitherFromHex(LP_USER), 10000n);
-				const nightCoin1 = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				const result = calculateAddLiquidityAmounts(
-					10000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin1,
-					nightCoin1,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get actual reserves
-				const [reserveUSDC, reserveNIGHT] = lunarswap.getPairReserves(
-					usdcCoin1,
-					nightCoin1,
-				);
-
-				// Calculate optimal amounts
-				const result2 = calculateAddLiquidityAmounts(
-					1000n, // desired USDC
-					10000n, // desired NIGHT
-					reserveUSDC, // actual reserve USDC
-					reserveNIGHT, // actual reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-
-				// Try to add liquidity with a very different ratio
-				const usdcCoin2 = usdc.mint(createEitherFromHex(LP_USER), 1000n);
-				const nightCoin2 = night.mint(createEitherFromHex(LP_USER), 10000n);
-
-				try {
-					lunarswap.addLiquidity(
-						usdcCoin2,
-						nightCoin2,
-						result2.amountAMin, // calculated minimum USDC
-						5000n, // amountBMin much higher than optimal (should be ~99)
-						recipient,
-					);
-					// If no error is thrown, fail the test
-					throw new Error("Expected error was not thrown");
-				} catch (e: unknown) {
-					let message = "";
-					if (e instanceof Error) {
-						message = e.message;
-					} else if (typeof e === "string") {
-						message = e;
-					}
-					// Accept either Insufficient A or B amount error since token order can vary
-					expect(
-						message.includes("LunarswapRouter: Insufficient A amount") ||
-							message.includes("LunarswapRouter: Insufficient B amount"),
-					).toBe(true);
-				}
-			});
-
-			/**
-			 * Tests multiple rapid liquidity additions
-			 *
-			 * Mathematical calculations:
-			 * - First: 2000 USDC, 2000 NIGHT → liquidity = sqrt(4,000,000) - 1000 = 2000 - 1000 = 1000
-			 * - Second: 2000 USDC, 2000 NIGHT → new liquidity = min((2000 * 1000) / 2000, (2000 * 1000) / 2000) = 1000
-			 * - Third: 2000 USDC, 2000 NIGHT → new liquidity = min((2000 * 2000) / 4000, (2000 * 2000) / 4000) = 1000
-			 * - Total liquidity = 1000 + 1000 + 1000 = 3000
-			 *
-			 * Expected: Accumulated liquidity > 1000, reserves = 6000 USDC, 6000 NIGHT
-			 */
-			it("should handle multiple rapid liquidity additions", () => {
-				const recipient = createEitherFromHex(LP_USER);
-
-				// First addition
-				const usdcCoin1 = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin1 = night.mint(createEitherFromHex(LP_USER), 2000n);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					2000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin1,
-					nightCoin1,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Second addition - calculate optimal amounts
-				const usdcCoin2 = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin2 = night.mint(createEitherFromHex(LP_USER), 2000n);
-				const [reserveUSDC2, reserveNIGHT2] = lunarswap.getPairReserves(
-					usdcCoin1,
-					nightCoin1,
-				);
-
-				const result2 = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					2000n, // desired NIGHT
-					reserveUSDC2, // reserve USDC
-					reserveNIGHT2, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-
-				lunarswap.addLiquidity(
-					usdcCoin2,
-					nightCoin2,
-					result2.amountAMin,
-					result2.amountBMin,
-					recipient,
-				);
-
-				// Third addition - calculate optimal amounts
-				const usdcCoin3 = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin3 = night.mint(createEitherFromHex(LP_USER), 2000n);
-				const [reserveUSDC3, reserveNIGHT3] = lunarswap.getPairReserves(
-					usdcCoin1,
-					nightCoin1,
-				);
-
-				const result3 = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					2000n, // desired NIGHT
-					reserveUSDC3, // reserve USDC
-					reserveNIGHT3, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin3,
-					nightCoin3,
-					result3.amountAMin,
-					result3.amountBMin,
-					recipient,
-				);
-
-				// Verify the pair has accumulated liquidity
-				expect(lunarswap.getLpTokenTotalSupply(usdcCoin1, nightCoin1).value).toBe(6000n); // multiple rapid additions
-			});
-
-			/**
-			 * Tests concurrent pair creation
-			 *
-			 * Mathematical calculations:
-			 * - Pair 1: 2000 USDC, 2000 NIGHT → liquidity = sqrt(4,000,000) - 1000 = 1000
-			 * - Pair 2: 2000 USDC, 2000 DUST → liquidity = sqrt(4,000,000) - 1000 = 1000
-			 * - Pair 3: 2000 NIGHT, 2000 DUST → liquidity = sqrt(4,000,000) - 1000 = 1000
-			 * - Each pair gets 1000 liquidity tokens
-			 *
-			 * Expected: 3 unique pairs created, total pairs length = 3
-			 */
-			it("should handle concurrent pair creation", () => {
-				const recipient = createEitherFromHex(LP_USER);
-
-				// Create multiple pairs simultaneously
-				const usdcCoin1 = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin1 = night.mint(createEitherFromHex(LP_USER), 2000n);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					2000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin1,
-					nightCoin1,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				const usdcCoin2 = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const dustCoin1 = dust.mint(createEitherFromHex(LP_USER), 2000n);
-				const result2 = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					2000n, // desired DUST
-					0n, // reserve USDC
-					0n, // reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin2,
-					dustCoin1,
-					result2.amountAMin,
-					result2.amountBMin,
-					recipient,
-				);
-
-				const nightCoin2 = night.mint(createEitherFromHex(LP_USER), 2000n);
-				const dustCoin2 = dust.mint(createEitherFromHex(LP_USER), 2000n);
-				const result3 = calculateAddLiquidityAmounts(
-					2000n, // desired NIGHT
-					2000n, // desired DUST
-					0n, // reserve NIGHT
-					0n, // reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					nightCoin2,
-					dustCoin2,
-					result3.amountAMin,
-					result3.amountBMin,
-					recipient,
-				);
-
-				expect(lunarswap.getAllPairLength()).toBe(3n);
-			});
-
-			/**
-			 * Tests minimum viable amounts for liquidity provision
-			 *
-			 * Mathematical calculations:
-			 * - Input: 2000 USDC, 2000 NIGHT
-			 * - Liquidity = sqrt(2000 * 2000) - MINIMUM_LIQUIDITY(1000)
-			 * - Liquidity = sqrt(4,000,000) - 1000 = 2000 - 1000 = 1000
-			 * - Min amounts: 1800 USDC, 1800 NIGHT (90% of input)
-			 *
-			 * Expected: Successful liquidity provision with 1000 liquidity tokens
-			 */
-			it("should handle edge case with minimum viable amounts", () => {
-				// Test with amounts that are just above the minimum liquidity threshold
-				const minViableAmount = 2000n; // Increased to ensure sufficient liquidity
-				const usdcCoin = usdc.mint(
-					createEitherFromHex(LP_USER),
-					minViableAmount,
-				);
-				const nightCoin = night.mint(
-					createEitherFromHex(LP_USER),
-					minViableAmount,
-				);
-				const recipient = createEitherFromHex(LP_USER);
-
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					2000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Use getPairIdentity to get the correct order for getIdentity
-				expect(lunarswap.getLpTokenTotalSupply(usdcCoin, nightCoin).value).toBe(2000n);
-			});
-		});
-	});
-
-	describe("removeLiquidity", () => {
-		describe("USDC/NIGHT pair", () => {
-			/**
-			 * Tests removing liquidity from a USDC/NIGHT pair
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - Total LP supply = 1414 (MINIMUM_LIQUIDITY + liquidity)
-			 * - Remove all burnable LP tokens: 414 LP tokens
-			 * - Expected token0 amount = (414 * 2000) / 1414 = 585 (integer division)
-			 * - Expected token1 amount = (414 * 1000) / 1414 = 292 (integer division)
-			 * - Remaining reserves: 1415 USDC (2000 - 585), 708 NIGHT (1000 - 292)
-			 * - Remaining LP tokens: 1000 (MINIMUM_LIQUIDITY)
-			 *
-			 * Expected: Correct proportional token amounts returned, reserves updated
-			 */
-			it("should remove liquidity from USDC/NIGHT pair", () => {
-				// First, add liquidity to create the pair and get LP tokens
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Create LP token coin for removal (use all burnable liquidity)
-				const lpTokensToRemove = initialLpTotalSupply.value - 1000n;
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: lpTokensToRemove,
-				};
-
-				// Calculate minimum amounts using SDK
-				const [reserveA, reserveB] = lunarswap.getPairReserves(
-					usdcCoin,
-					nightCoin,
-				);
-				const { amountAMin, amountBMin } = calculateRemoveLiquidityMinimums(
-					lpTokensToRemove,
-					initialLpTotalSupply.value,
-					reserveA,
-					reserveB,
-					SLIPPAGE_TOLERANCE.LOW,
-				);
-
-				// Remove liquidity
-				lunarswap.removeLiquidity(
-					usdcCoin,
-					nightCoin,
-					lpTokenCoin,
-					amountAMin,
-					amountBMin,
-					recipient,
-				);
-
-				// Get updated pair state
-				const updatedPair = lunarswap.getPair(usdcCoin, nightCoin);
-				const updatedLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-				const [updatedReserveA, updatedReserveB] = lunarswap.getPairReserves(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Verify LP token supply decreased
-				expect(updatedLpTotalSupply.value).toBe(
-					initialLpTotalSupply.value - lpTokensToRemove,
-				);
-
-				// Verify reserves decreased proportionally
-				const expectedValues = getExpectedTokenValues(
-					updatedPair,
-					usdcCoin,
-					nightCoin,
-					1415n, // Correct reserve for USDC after removal (2000 - 585)
-					708n, // Correct reserve for NIGHT after removal (1000 - 292)
-					lunarswap,
-				);
-				expect(updatedPair.token0.value).toBe(expectedValues.token0Value);
-				expect(updatedPair.token1.value).toBe(expectedValues.token1Value);
-			});
-
-			/**
-			 * Tests removing all burnable liquidity from a USDC/NIGHT pair
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - MINIMUM_LIQUIDITY = 1000, total LP supply = 1414
-			 * - Remove all burnable LP tokens: 414
-			 * - Remaining LP tokens: 1000 (MINIMUM_LIQUIDITY)
-			 * - Expected reserves (integer division):
-			 *   USDC: (1000 * 2000) / 1414 = 1415
-			 *   NIGHT: (1000 * 1000) / 1414 = 708
-			 *
-			 * Expected: All removable liquidity removed, MINIMUM_LIQUIDITY remains, reserves match proportionally
-			 */
-			it("should remove all removable liquidity from USDC/NIGHT pair", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Remove all LP tokens except MINIMUM_LIQUIDITY (1000)
-				const removableLpTokens = initialLpTotalSupply.value - 1000n;
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: removableLpTokens,
-				};
-
-				// Calculate minimum amounts using SDK
-				const [reserveA, reserveB] = lunarswap.getPairReserves(
-					usdcCoin,
-					nightCoin,
-				);
-				const { amountAMin, amountBMin } = calculateRemoveLiquidityMinimums(
-					removableLpTokens,
-					initialLpTotalSupply.value,
-					reserveA,
-					reserveB,
-					SLIPPAGE_TOLERANCE.LOW,
-				);
-
-				// Remove liquidity
-				lunarswap.removeLiquidity(
-					usdcCoin,
-					nightCoin,
-					lpTokenCoin,
-					amountAMin,
-					amountBMin,
-					recipient,
-				);
-
-				// Get updated pair state
-				const updatedPair = lunarswap.getPair(usdcCoin, nightCoin);
-				const updatedLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Verify only MINIMUM_LIQUIDITY remains
-				expect(updatedLpTotalSupply.value).toBe(1000n);
-
-				// Verify reserves are reduced but not zero
-				const expectedValues = getExpectedTokenValues(
-					updatedPair,
-					usdcCoin,
-					nightCoin,
-					1415n, // MINIMUM_LIQUIDITY worth of USDC
-					708n, // MINIMUM_LIQUIDITY worth of NIGHT
-					lunarswap,
-				);
-				expect(updatedPair.token0.value).toBe(expectedValues.token0Value);
-				expect(updatedPair.token1.value).toBe(expectedValues.token1Value);
-			});
-
-			/**
-			 * Tests removing small amounts of liquidity from USDC/NIGHT pair
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - Remove 10% of LP tokens: 141 LP tokens
-			 * - Expected token0 amount = (141 * 2000) / 1414 = 199 (integer division)
-			 * - Expected token1 amount = (141 * 1000) / 1414 = 99 (integer division)
-			 * - Remaining reserves: 1801 USDC (2000 - 199), 901 NIGHT (1000 - 99)
-			 *
-			 * Expected: Small proportional amounts returned, reserves updated correctly
-			 */
-			it("should remove small amounts of liquidity from USDC/NIGHT pair", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Remove 10% of LP tokens
-				const lpTokensToRemove = initialLpTotalSupply.value / 10n;
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: lpTokensToRemove,
-				};
-
-				// Calculate minimum amounts using SDK
-				const [reserveA, reserveB] = lunarswap.getPairReserves(
-					usdcCoin,
-					nightCoin,
-				);
-				const { amountAMin, amountBMin } = calculateRemoveLiquidityMinimums(
-					lpTokensToRemove,
-					initialLpTotalSupply.value,
-					reserveA,
-					reserveB,
-					SLIPPAGE_TOLERANCE.LOW,
-				);
-
-				// Remove liquidity
-				lunarswap.removeLiquidity(
-					usdcCoin,
-					nightCoin,
-					lpTokenCoin,
-					amountAMin,
-					amountBMin,
-					recipient,
-				);
-
-				// Get updated pair state
-				const updatedPair = lunarswap.getPair(usdcCoin, nightCoin);
-				const updatedLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Verify LP token supply decreased
-				expect(updatedLpTotalSupply.value).toBe(
-					initialLpTotalSupply.value - lpTokensToRemove,
-				);
-
-				// Verify reserves decreased proportionally
-				const expectedValues = getExpectedTokenValues(
-					updatedPair,
-					usdcCoin,
-					nightCoin,
-					1801n, // 2000 - 199
-					901n, // 1000 - 99
-					lunarswap,
-				);
-				expect(updatedPair.token0.value).toBe(expectedValues.token0Value);
-				expect(updatedPair.token1.value).toBe(expectedValues.token1Value);
-			});
-		});
-
-		describe("NIGHT/DUST pair", () => {
-			/**
-			 * Tests removing liquidity from a NIGHT/DUST pair
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 8000 NIGHT, 12000 DUST → liquidity = 8798
-			 * - Remove 30% of LP tokens: 2939 LP tokens
-			 * - Expected token0 amount = (2939 * 8000) / 9798 = 2399 (integer division)
-			 * - Expected token1 amount = (2939 * 12000) / 9798 = 3599 (integer division)
-			 * - Remaining reserves: 5601 NIGHT (8000 - 2399), 8401 DUST (12000 - 3599)
-			 *
-			 * Expected: Correct proportional token amounts returned, reserves updated
-			 */
-			it("should remove liquidity from NIGHT/DUST pair", () => {
-				// First, add liquidity to create the pair
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 8000n);
-				const dustCoin = dust.mint(createEitherFromHex(LP_USER), 12000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					8000n, // desired NIGHT
-					12000n, // desired DUST
-					0n, // reserve NIGHT
-					0n, // reserve DUST
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					nightCoin,
-					dustCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const pair = lunarswap.getPair(nightCoin, dustCoin);
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					nightCoin,
-					dustCoin,
-				);
-
-				// Remove 30% of LP tokens
-				const lpTokensToRemove = (initialLpTotalSupply.value * 3n) / 10n;
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: lpTokensToRemove,
-				};
-
-				// Calculate minimum amounts using SDK
-				const [reserveA, reserveB] = lunarswap.getPairReserves(
-					nightCoin,
-					dustCoin,
-				);
-				const { amountAMin, amountBMin } = calculateRemoveLiquidityMinimums(
-					lpTokensToRemove,
-					initialLpTotalSupply.value,
-					reserveA,
-					reserveB,
-					SLIPPAGE_TOLERANCE.LOW,
-				);
-
-				// Remove liquidity
-				lunarswap.removeLiquidity(
-					nightCoin,
-					dustCoin,
-					lpTokenCoin,
-					amountAMin,
-					amountBMin,
-					recipient,
-				);
-
-				// Get updated pair state
-				const updatedPair = lunarswap.getPair(nightCoin, dustCoin);
-				const updatedLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					nightCoin,
-					dustCoin,
-				);
-
-				// Verify LP token supply decreased
-				expect(updatedLpTotalSupply.value).toBe(
-					initialLpTotalSupply.value - lpTokensToRemove,
-				);
-
-				// Verify reserves decreased proportionally
-				const expectedValues = getExpectedTokenValues(
-					updatedPair,
-					nightCoin,
-					dustCoin,
-					5601n, // 8000 - 2399
-					8401n, // 12000 - 3599
-					lunarswap,
-				);
-				expect(updatedPair.token0.value).toBe(expectedValues.token0Value);
-				expect(updatedPair.token1.value).toBe(expectedValues.token1Value);
-			});
-		});
-
-		describe("edge cases", () => {
-			/**
-			 * Tests removing liquidity with insufficient minimum amounts
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - Total LP supply = 1414 (MINIMUM_LIQUIDITY + liquidity)
-			 * - Remove 50% of LP tokens: 707 LP tokens
-			 * - Expected token0 amount = (707 * 2000) / 1414 = 1000 (integer division)
-			 * - Expected token1 amount = (707 * 1000) / 1414 = 500 (integer division)
-			 * - Set amountAMin = 300 < 1000, should pass
-			 * - Set amountBMin = 150 < 500, should pass
-			 * - But if we set amountAMin = 1100 > 1000, should fail
-			 *
-			 * Error: "LunarswapLibrary: subQualifiedCoinValue() - Insufficient amount" - minimum not met
-			 */
-			it("should fail when removing liquidity with insufficient minimum amounts", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Remove 50% of LP tokens
-				const lpTokensToRemove = initialLpTotalSupply.value / 2n;
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: lpTokensToRemove,
-				};
-
-				// Try to remove liquidity with minimum amounts higher than what will be returned
-				expect(() => {
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						lpTokenCoin,
-						1100n, // amountAMin higher than expected (~1000)
-						600n, // amountBMin higher than expected (~500)
-						recipient,
-					);
-				}).toThrow(
-					"LunarswapRouter: Insufficient A amount",
-				);
-			});
-
-			/**
-			 * Tests removing liquidity with zero LP tokens
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - Try to remove 0 LP tokens
-			 * - Expected token0 amount = (0 * 2000) / 1414 = 0 (integer division)
-			 * - Expected token1 amount = (0 * 1000) / 1414 = 0 (integer division)
-			 * - Since amounts are 0, "Insufficient liquidity burned" error
-			 *
-			 * Error: "LunarswapPair: burn() - Insufficient liquidity burned" - zero amounts
-			 */
-			it("should fail when removing liquidity with zero LP tokens", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Try to remove 0 LP tokens
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: 0n,
-				};
-
-				expect(() => {
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						lpTokenCoin,
-						0n, // amountAMin
-						0n, // amountBMin
-						recipient,
-					);
-				}).toThrow("LunarswapPair: burn() - Insufficient liquidity burned");
-			});
-
-			/**
-			 * Tests removing liquidity with LP tokens exceeding total supply
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - Total LP supply = 1414
-			 * - Try to remove 2000 LP tokens > 1414
-			 * - This should fail due to insufficient LP tokens
-			 *
-			 * Error: Should fail due to insufficient LP tokens for burning
-			 */
-			it("should fail when removing liquidity with LP tokens exceeding total supply", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Try to remove more LP tokens than exist
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: initialLpTotalSupply.value + 1000n, // More than total supply
-				};
-
-				expect(() => {
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						lpTokenCoin,
-						1000n, // amountAMin
-						500n, // amountBMin
-						recipient,
-					);
-				}).toThrow("LunarswapPair: burn() - Insufficient reserves for token0");
-			});
-
-			/**
-			 * Tests removing liquidity from non-existent pair
-			 *
-			 * Mathematical validation:
-			 * - No pair exists, so no LP tokens exist
-			 * - Trying to remove liquidity from non-existent pair should fail
-			 * - LP token total supply lookup will fail
-			 *
-			 * Error: "LunarswapLpTokens: totalSupply() - Lp token not found"
-			 */
-			it("should fail when removing liquidity from non-existent pair", () => {
-				// Create tokens but don't add liquidity
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				// Create fake LP token coin
-				const fakeLpTokenCoin = {
-					nonce: new Uint8Array(32).fill(0x01),
-					color: new Uint8Array(32).fill(0x02),
-					value: 100n,
-				};
-
-				expect(() => {
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						fakeLpTokenCoin,
-						50n, // amountAMin
-						50n, // amountBMin
-						recipient,
-					);
-				}).toThrow("LunarswapFactory: getPair() - Pair does not exist");
-			});
-
-			/**
-			 * Tests removing liquidity with very small amounts
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - Remove 1 LP token
-			 * - Expected token0 amount = (1 * 2000) / 1414 = 1 (integer division)
-			 * - Expected token1 amount = (1 * 1000) / 1414 = 0 (integer division)
-			 * - Very small amounts may round down to 0, causing "Insufficient liquidity burned"
-			 *
-			 * Expected: Should handle very small amounts or fail gracefully
-			 */
-			it("should handle removing very small amounts of liquidity", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Try to remove 1 LP token
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: 1n,
-				};
-
-				try {
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						lpTokenCoin,
-						0n, // amountAMin
-						0n, // amountBMin
-						recipient,
-					);
-					// If successful, verify the change
-					const updatedLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-						usdcCoin,
-						nightCoin,
-					);
-					expect(updatedLpTotalSupply.value).toBeGreaterThan(0n);
-				} catch (e: unknown) {
-					// If it fails, it should be due to insufficient liquidity burned
-					let message = "";
-					if (e instanceof Error) {
-						message = e.message;
-					} else if (typeof e === "string") {
-						message = e;
-					}
-					expect(message).toContain("Insufficient liquidity burned");
-				}
-			});
-
-			/**
-			 * Tests removing liquidity multiple times from same pair
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - First removal: 20% of LP tokens (283 LP tokens)
-			 *   Expected token0 = (283 * 2000) / 1414 = 400 (integer division)
-			 *   Expected token1 = (283 * 1000) / 1414 = 200 (integer division)
-			 * - Second removal: 30% of remaining LP tokens (339 LP tokens)
-			 *   Expected token0 = (339 * 1600) / 1131 = 479 (integer division)
-			 *   Expected token1 = (339 * 800) / 1131 = 239 (integer division)
-			 * - Third removal: 50% of remaining LP tokens (396 LP tokens)
-			 *   Expected token0 = (396 * 1121) / 792 = 560 (integer division)
-			 *   Expected token1 = (396 * 561) / 792 = 280 (integer division)
-			 * - Final LP tokens: 396 remaining
-			 *
-			 * Expected: Multiple removals work correctly, reserves updated properly
-			 */
-			it("should handle multiple liquidity removals from same pair", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const pair = lunarswap.getPair(usdcCoin, nightCoin);
-				let currentLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				try {
-					// First removal: 20% of LP tokens
-					const firstRemoval = currentLpTotalSupply.value / 5n;
-					const lpTokenCoin1 = {
-						nonce: currentLpTotalSupply.nonce,
-						color: currentLpTotalSupply.color,
-						value: firstRemoval,
-					};
-
-					// Calculate minimum amounts using SDK for first removal
-					const [reserveA1, reserveB1] = lunarswap.getPairReserves(
-						usdcCoin,
-						nightCoin,
-					);
-					const { amountAMin: amountAMin1, amountBMin: amountBMin1 } =
-						calculateRemoveLiquidityMinimums(
-							firstRemoval,
-							currentLpTotalSupply.value,
-							reserveA1,
-							reserveB1,
-							SLIPPAGE_TOLERANCE.LOW,
-						);
-
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						lpTokenCoin1,
-						amountAMin1,
-						amountBMin1,
-						recipient,
-					);
-
-					currentLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-						usdcCoin,
-						nightCoin,
-					);
-
-					// Second removal: 30% of remaining LP tokens
-					const secondRemoval = (currentLpTotalSupply.value * 3n) / 10n;
-					const lpTokenCoin2 = {
-						nonce: currentLpTotalSupply.nonce,
-						color: currentLpTotalSupply.color,
-						value: secondRemoval,
-					};
-
-					// Calculate minimum amounts using SDK for second removal
-					const [reserveA2, reserveB2] = lunarswap.getPairReserves(
-						usdcCoin,
-						nightCoin,
-					);
-					const { amountAMin: amountAMin2, amountBMin: amountBMin2 } =
-						calculateRemoveLiquidityMinimums(
-							secondRemoval,
-							currentLpTotalSupply.value,
-							reserveA2,
-							reserveB2,
-							SLIPPAGE_TOLERANCE.LOW,
-						);
-
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						lpTokenCoin2,
-						amountAMin2,
-						amountBMin2,
-						recipient,
-					);
-
-					currentLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-						usdcCoin,
-						nightCoin,
-					);
-
-					// Third removal: 50% of remaining LP tokens
-					const thirdRemoval = currentLpTotalSupply.value / 2n;
-					const lpTokenCoin3 = {
-						nonce: currentLpTotalSupply.nonce,
-						color: currentLpTotalSupply.color,
-						value: thirdRemoval,
-					};
-
-					// Calculate minimum amounts using SDK for third removal
-					const [reserveA3, reserveB3] = lunarswap.getPairReserves(
-						usdcCoin,
-						nightCoin,
-					);
-					const { amountAMin: amountAMin3, amountBMin: amountBMin3 } =
-						calculateRemoveLiquidityMinimums(
-							thirdRemoval,
-							currentLpTotalSupply.value,
-							reserveA3,
-							reserveB3,
-							SLIPPAGE_TOLERANCE.LOW,
-						);
-
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						lpTokenCoin3,
-						amountAMin3,
-						amountBMin3,
-						recipient,
-					);
-
-					// Verify final state
-					const finalLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-						usdcCoin,
-						nightCoin,
-					);
-					const finalPair = lunarswap.getPair(usdcCoin, nightCoin);
-
-					// Should have some LP tokens remaining
-					expect(finalLpTotalSupply.value).toBeGreaterThan(0n);
-					expect(finalPair.token0.value).toBeGreaterThan(0n);
-					expect(finalPair.token1.value).toBeGreaterThan(0n);
-				} catch (e: unknown) {
-					// If it fails, it should be due to insufficient liquidity
-					let message = "";
-					if (e instanceof Error) {
-						message = e.message;
-					} else if (typeof e === "string") {
-						message = e;
-					}
-					expect(message).toContain(
-						"LunarswapLibrary: subQualifiedCoinValue() - Insufficient amount",
-					);
-				}
-			});
-		});
-
-		describe("error handling", () => {
-			/**
-			 * Tests removing liquidity with invalid LP token nonce
-			 *
-			 * Mathematical validation:
-			 * - LP token has wrong nonce that doesn't match the pair
-			 * - This should fail during LP token validation
-			 *
-			 * Error: Should fail due to invalid LP token nonce
-			 */
-			it("should fail when removing liquidity with invalid LP token nonce", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Create LP token with completely wrong nonce and color
-				const invalidLpTokenCoin = {
-					nonce: new Uint8Array(32).fill(0x99), // Wrong nonce
-					color: new Uint8Array(32).fill(0x88), // Wrong color
-					value: 100n,
-				};
-
-				// This should throw an error due to invalid LP token
-				expect(() => {
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						invalidLpTokenCoin,
-						50n,
-						50n,
-						recipient,
-					);
-				}).toThrow("LunarswapRouter: mismatched LP token color");
-			});
-
-			/**
-			 * Tests removing liquidity with wrong token order
-			 *
-			 * Mathematical validation:
-			 * - Tokens are passed in wrong order (NIGHT, USDC instead of USDC, NIGHT)
-			 * - This should still work as tokens are sorted internally
-			 *
-			 * Expected: Should work correctly despite wrong order
-			 */
-			it("should handle removing liquidity with wrong token order", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Remove 5% of LP tokens with wrong token order
-				const lpTokensToRemove = initialLpTotalSupply.value / 20n;
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: lpTokensToRemove,
-				};
-
-				// Remove liquidity with wrong token order (nightCoin, usdcCoin) and zero minimums
-				lunarswap.removeLiquidity(
-					nightCoin, // Wrong order
-					usdcCoin,  // Wrong order
-					lpTokenCoin,
-					0n, // amountAMin - use zero to avoid insufficient amount error
-					0n, // amountBMin - use zero to avoid insufficient amount error
-					recipient,
-				);
-
-				// Verify the operation succeeded
-				const updatedLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-				expect(updatedLpTotalSupply.value).toBe(
-					initialLpTotalSupply.value - lpTokensToRemove,
-				);
-			});
-
-			/**
-			 * Tests removing liquidity with identical token addresses
-			 *
-			 * Mathematical validation:
-			 * - Both tokens have same color (USDC = USDC)
-			 * - This should fail during token validation
-			 *
-			 * Error: "Lunarswap: removeLiquidity() - Identical addresses"
-			 */
-			it("should fail when removing liquidity with identical token addresses", () => {
-				// Create tokens but don't add liquidity
-				const usdcCoin1 = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const usdcCoin2 = usdc.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-
-				// Create fake LP token coin
-				const fakeLpTokenCoin = {
-					nonce: new Uint8Array(32).fill(0x01),
-					color: new Uint8Array(32).fill(0x02),
-					value: 100n,
-				};
-
-				expect(() => {
-					lunarswap.removeLiquidity(
-						usdcCoin1,
-						usdcCoin2, // Same token type
-						fakeLpTokenCoin,
-						50n, // amountAMin
-						50n, // amountBMin
-						recipient,
-					);
-				}).toThrow("Lunarswap: removeLiquidity() - Identical addresses");
-			});
-
-			/**
-			 * Tests removing liquidity with negative minimum amounts
-			 *
-			 * Mathematical validation:
-			 * - Negative amounts should be rejected
-			 * - This should fail during parameter validation
-			 *
-			 * Error: Should fail due to negative amounts
-			 */
-			it("should fail when removing liquidity with negative minimum amounts", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Create LP token coin
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: 100n,
-				};
-
-				// Try to remove liquidity with negative minimum amounts
-				// Note: In bigint, negative values are represented differently
-				// This test checks if the system handles edge cases properly
-				expect(() => {
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						lpTokenCoin,
-						0n, // amountAMin (minimum allowed)
-						0n, // amountBMin (minimum allowed)
-						recipient,
-					);
-				}).not.toThrow(); // Should work with zero minimums
-			});
-
-			/**
-			 * Tests removing liquidity with excessive minimum amounts
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - Remove 10% of LP tokens: 141 LP tokens
-			 * - Expected token0 amount = (141 * 2000) / 1414 = 199 (integer division)
-			 * - Expected token1 amount = (141 * 1000) / 1414 = 99 (integer division)
-			 * - Set amountAMin = 1000 > 199, should fail
-			 *
-			 * Error: "LunarswapRouter: Insufficient A amount" - minimum too high
-			 */
-			it("should fail when removing liquidity with excessive minimum amounts", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Remove 10% of LP tokens
-				const lpTokensToRemove = initialLpTotalSupply.value / 10n;
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: lpTokensToRemove,
-				};
-
-				// Try to remove liquidity with excessive minimum amounts
-				expect(() => {
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						lpTokenCoin,
-						1000n, // amountAMin much higher than expected (~199)
-						200n,  // amountBMin much higher than expected (~99)
-						recipient,
-					);
-				}).toThrow("LunarswapRouter: Insufficient A amount");
-			});
-
-			/**
-			 * Tests removing liquidity with mismatched LP token color
-			 *
-			 * Mathematical validation:
-			 * - LP token has wrong color that doesn't match the pair
-			 * - This should fail during LP token validation
-			 *
-			 * Error: Should fail due to mismatched LP token color
-			 */
-			it("should fail when removing liquidity with mismatched LP token color", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Create LP token with wrong color but correct nonce
-				const mismatchedLpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce, // Correct nonce
-					color: new Uint8Array(32).fill(0xAA), // Wrong color
-					value: 100n,
-				};
-
-				// This should throw an error due to mismatched LP token color
-				expect(() => {
-					lunarswap.removeLiquidity(
-						usdcCoin,
-						nightCoin,
-						mismatchedLpTokenCoin,
-						50n,
-						50n,
-						recipient,
-					);
-				}).toThrow("LunarswapRouter: mismatched LP token color");
-			});
-
-			/**
-			 * Tests removing liquidity with zero minimum amounts
-			 *
-			 * Mathematical calculations:
-			 * - Initial: 2000 USDC, 1000 NIGHT → liquidity = 414
-			 * - Remove 5% of LP tokens: 71 LP tokens
-			 * - Expected token0 amount = (71 * 2000) / 1414 = 100 (integer division)
-			 * - Expected token1 amount = (71 * 1000) / 1414 = 50 (integer division)
-			 * - Set amountAMin = 0, amountBMin = 0, should pass
-			 *
-			 * Expected: Should work with zero minimum amounts
-			 */
-			it("should handle removing liquidity with zero minimum amounts", () => {
-				// First, add liquidity to create the pair
-				const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 2000n);
-				const nightCoin = night.mint(createEitherFromHex(LP_USER), 1000n);
-				const recipient = createEitherFromHex(LP_USER);
-				const result = calculateAddLiquidityAmounts(
-					2000n, // desired USDC
-					1000n, // desired NIGHT
-					0n, // reserve USDC
-					0n, // reserve NIGHT
-					SLIPPAGE_TOLERANCE.LOW, // 0.5% slippage
-				);
-				lunarswap.addLiquidity(
-					usdcCoin,
-					nightCoin,
-					result.amountAMin,
-					result.amountBMin,
-					recipient,
-				);
-
-				// Get initial pair state
-				const initialLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-
-				// Remove 5% of LP tokens
-				const lpTokensToRemove = initialLpTotalSupply.value / 20n;
-				const lpTokenCoin = {
-					nonce: initialLpTotalSupply.nonce,
-					color: initialLpTotalSupply.color,
-					value: lpTokensToRemove,
-				};
-
-				// Remove liquidity with zero minimum amounts
-				lunarswap.removeLiquidity(
-					usdcCoin,
-					nightCoin,
-					lpTokenCoin,
-					0n, // amountAMin
-					0n, // amountBMin
-					recipient,
-				);
-
-				// Verify the operation succeeded
-				const updatedLpTotalSupply = lunarswap.getLpTokenTotalSupply(
-					usdcCoin,
-					nightCoin,
-				);
-				expect(updatedLpTotalSupply.value).toBe(
-					initialLpTotalSupply.value - lpTokensToRemove,
-				);
-			});
-		});
-	});
-
 	describe("isPairExists", () => {
 		it("should return false for non-existent pair", () => {
 			const usdcCoin = usdc.mint(createEitherFromHex(LP_USER), 100n);
@@ -2768,7 +329,9 @@ describe("Lunarswap", () => {
 			);
 			expect(pair.token0.value).toBe(expectedValues.token0Value);
 			expect(pair.token1.value).toBe(expectedValues.token1Value);
-			expect(lunarswap.getLpTokenTotalSupply(usdcCoin, dustCoin).value).toBe(14142n);
+			expect(lunarswap.getLpTokenTotalSupply(usdcCoin, dustCoin).value).toBe(
+				14142n,
+			);
 			expect(pair.kLast).toBe(0n);
 		});
 
@@ -2809,7 +372,9 @@ describe("Lunarswap", () => {
 			);
 			expect(pair.token0.value).toBe(expectedValues.token0Value);
 			expect(pair.token1.value).toBe(expectedValues.token1Value);
-			expect(lunarswap.getLpTokenTotalSupply(nightCoin, dustCoin).value).toBe(9797n);
+			expect(lunarswap.getLpTokenTotalSupply(nightCoin, dustCoin).value).toBe(
+				9797n,
+			);
 			expect(pair.kLast).toBe(0n);
 		});
 	});
@@ -2943,7 +508,9 @@ describe("Lunarswap", () => {
 
 			// Use getPairIdentity to get the correct order for getLpTokenTotalSupply
 			const identity = lunarswap.getPairIdentity(usdcCoin, nightCoin);
-			expect(lunarswap.getLpTokenTotalSupply(usdcCoin, nightCoin).value).toBe(7071n); // LP token total supply tracking
+			expect(lunarswap.getLpTokenTotalSupply(usdcCoin, nightCoin).value).toBe(
+				7071n,
+			); // LP token total supply tracking
 		});
 	});
 });
